Day 17 - True or False Game
This project is used to show how to use OOP in python. All classes were made from scratch and allow the user to use a set of data to ask a player a set of true/false questions. It also works with long answer
questions if the player inputs the exact answer correctly.