Skipping Day 39-40.
Unfortunately, I will  be moving on from day 39, and 40 as well since they are connected. I have been working on 39 on and off for a few weeks now. Due to the complexity of the project, and my little time, I think it is fair to just move on and work on this another time. I still have 60 days, so I am not worried. I will come back.